+++ docs/flutter_bootstrap.js
@@ -38,6 +38,6 @@ _flutter.buildConfig = {"engineRevision":"a8bfdfc394deaed5c57bd45a64ac4294dc976a
 
 _flutter.loader.load({
   serviceWorkerSettings: {
-    serviceWorkerVersion: "2431400873"
+    serviceWorkerVersion: "3582425447"
   }
 });